requirements.txt
flask==3.0.0
langchain==0.1.0
langchain-openai==0.0.5
chromadb==0.4.22
openai==1.12.0
PyMuPDF==1.23.0
python-dotenv==1.0.0
pypdf==3.17.0
tiktoken==0.5.2
